requirements.txt
plaidml==0.7.0
plaidml-keras==0.7.0
plaidbench==0.7.0
Pillow==7.2.0
pydot==1.4.1
matplotlib==3.3.0
numpy==1.19.1
tensorflow==1.15.3
tensorboard==1.15.0